Moved errors files
