+++ src/meilisearch.ts
@@ -11,7 +11,8 @@ import { Index } from './index'
 import MeiliAxiosWrapper from './meili-axios-wrapper'
 import * as Types from './types'
 
-class Meilisearch extends MeiliAxiosWrapper {
+class Meilisearch extends MeiliAxiosWrapper
+  implements Types.MeiliSearchInterface {
   config: Types.Config
   constructor(config: Types.Config) {
     super(config)
